+++ src/main/java/terramine/client/render/ModAllEntityRenderers.java
@@ -7,6 +7,7 @@ import net.fabricmc.fabric.api.event.client.ClientSpriteRegistryCallback;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.Sheets;
 import net.minecraft.client.renderer.entity.TntRenderer;
+import net.minecraft.world.inventory.InventoryMenu;
 import terramine.TerraMine;
 import terramine.client.render.entity.renderer.monsters.devourer.DevourerBodyRenderer;
 import terramine.client.render.entity.renderer.monsters.devourer.DevourerHeadRenderer;
@@ -25,6 +26,7 @@ import terramine.client.render.entity.renderer.projectiles.magic.RainbowMissileR
 import terramine.client.render.entity.renderer.projectiles.throwables.BombRenderer;
 import terramine.client.render.entity.renderer.projectiles.throwables.DynamiteRenderer;
 import terramine.client.render.entity.renderer.projectiles.throwables.GrenadeRenderer;
+import terramine.client.render.gui.TerrariaInventoryCreator;
 import terramine.common.init.ModBlockEntityType;
 import terramine.common.init.ModBlocks;
 import terramine.common.init.ModEntities;
@@ -74,8 +76,9 @@ public class ModAllEntityRenderers {
         EntityRendererRegistry.register(ModEntities.INSTANT_TNT, TntRenderer::new);
     }
 
+    // todo: remove and use atlas json in 1.19.3 and later
     private static void registerTextures() {
-        ClientSpriteRegistryCallback.event(Sheets.CHEST_SHEET).register((texture, registry) -> {
+        ClientSpriteRegistryCallback.event(Sheets.CHEST_SHEET).register((atlas, registry) -> {
             registry.register(TerraMine.id("block/chests/gold/gold_chest"));
             registry.register(TerraMine.id("block/chests/frozen/frozen_chest"));
             registry.register(TerraMine.id("block/chests/ivy/ivy_chest"));
@@ -85,6 +88,12 @@ public class ModAllEntityRenderers {
             registry.register(TerraMine.id("block/chests/shadow/shadow_chest"));
             registry.register(TerraMine.id("block/chests/player/piggy_bank/piggy_bank"));
             registry.register(TerraMine.id("block/chests/player/safe/safe"));
+        });
+
+        ClientSpriteRegistryCallback.event(InventoryMenu.BLOCK_ATLAS).register((atlas, registry) -> {
+            registry.register(TerrariaInventoryCreator.EMPTY_ACCESSORY_SLOT);
+            registry.register(TerrariaInventoryCreator.EMPTY_ACCESSORY_VANITY_SLOT);
+            registry.register(TerrariaInventoryCreator.EMPTY_ACCESSORY_DYE_SLOT);
         });
     }
 }