+++ src/main/java/terramine/mixin/player/InventoryScreenMixin.java
@@ -1,0 +1,41 @@
+package terramine.mixin.player;
+
+import net.minecraft.client.gui.components.ImageButton;
+import net.minecraft.client.gui.screens.inventory.EffectRenderingInventoryScreen;
+import net.minecraft.client.gui.screens.inventory.InventoryScreen;
+import net.minecraft.client.gui.screens.recipebook.RecipeUpdateListener;
+import net.minecraft.network.chat.Component;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.entity.player.Inventory;
+import net.minecraft.world.inventory.InventoryMenu;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.Unique;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import terramine.TerraMine;
+import terramine.client.render.gui.TerrariaInventoryHandler;
+
+@Mixin(InventoryScreen.class)
+public abstract class InventoryScreenMixin extends EffectRenderingInventoryScreen<InventoryMenu> implements RecipeUpdateListener {
+    @Unique
+    private static final ResourceLocation BUTTON_TEX = TerraMine.id("textures/gui/terraria_slots_button.png");
+
+    @Shadow
+    private boolean buttonClicked;
+
+    public InventoryScreenMixin(InventoryMenu abstractContainerMenu, Inventory inventory, Component component) {
+        super(abstractContainerMenu, inventory, component);
+    }
+
+    @Inject(method = "init", at = @At("TAIL"))
+    protected void onInit(CallbackInfo ci) {
+        if (!this.minecraft.gameMode.hasInfiniteItems()) {
+            this.addRenderableWidget(new ImageButton(this.leftPos + 66, this.height / 2 - 14, 8, 8, 0, 0, 8, BUTTON_TEX, 8, 16, (buttonWidget) -> {
+                this.minecraft.setScreen(new TerrariaInventoryHandler(this.minecraft.player));
+                this.buttonClicked = true;
+            }));
+        }
+    }
+}