+++ CHANGELOG.md
@@ -4,12 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.6.1-1.19] - 2022-12-01
-### Added
+## [1.7.0-1.19] - 2023-08-15
+// todo: update to 1.19.4 and then create a fork for 1.20
 - Meteorites // just need to make spawn in world then update can release
 - Meteorite armor, tools, and Space Gun
 - Hellstone ore
 - Hellstone armor and tools
+- New GUI // WIP
+- Some Dyes and Vanity accessories/armor // not started
+- Teams (will be used in the future) // not started
 
 ### Changed
 - Throwable explosives now render in 2D in Inventory